successful upload file from s3 and get the last element.
I use index requst and get only match Id to show picture.
and hide the left elelment of index using empty p and
make it inline so it not push the last element down to the button.
I need to fix the element by try to find the way to get just last index.
instead of hiding all index.